Deleting more (now) irrelevant parts of the game
We no longer render any HTML from the Go binary, so we can remove
home.go. We no longer use any WebSockets, so we can remove hub.go. We no
longer do whatever I was doing with passwords and cookies, so all that
garbage can go.

Make a mostly-complete database interface as banana.DB, and make a
simple in-memory implementation of it.