Start to add game creator stuff
The end goal is to allow only the person who made the game to hit 'Start
Game'. This requires having a concept of a user that exists beyond
joining the game, which requires pretty extensive changes actually.

The first couple changes are to introduce a new Register endpoint, and
start to introduce the logic for generating/signing/verifying tokens,
which the user will then send on all requests except Register.
Eventually, unauthenticated requests will result in redirects for the
end-user. I'm going with a JWT-based approach for now.

Also add a basic change to allow scaling the number of tiles for games
beyond 8 players, because I recently fonud out the hard way that games
go excruciatingly fast with 9 people. This change resulted in some basic
API cleaning up throughout the banana package.