Update the Docker image to specify an auth key path
Since the Docker image is pretty specific to my server anyway, continue
on that trend by specifying a path for the key directly in the Docker
image, so that the deployment understands auth tokens across resets.

The correct way to do this would be to use environment variables, but I
don't feel like it today.