Actually enforce the 'creator starts game' check
Addresses more of issue #8, now the person who created a game will be
the only one who can start it. Hide the button for people who didn't
create the game, since it won't work for them anyway.

This basically closes out the issue, but I'm going to keep it open to
track some work to make the flow nicer in the case that a user doesn't
exist (their cookie outlived the server) and stuff like that.