Finish refactor of account handling
Now, the auth interceptor will check if the account exists in the DB,
which means we can check on every endpoint if the account is in the DB,
and redirect accordingly.

Now, receiving an 'unauthenticated' error will result in being
redirected home, but auth is automatically performed if a name is set,
to speed up the process. Still not an ideal flow (ideally, auth would
automagically happen behind the scenes if needed, but c'est la vie).