Update the frontend to introduce the new min letters option
When creating a game, folks can now decide the minimum number of letters
required in a word for it to be valid. This doesn't actually work yet,
but the information is piped through everywhere now, so that's cool.

This CL also updates the frontend to change the layout a bit, adding
some instructions and some rambling.